Safety margins of multiple constraints
Maximize the minimum of a set of functions. Useful if unsafe set is basic semialgebraic {x | gi(x) >= 0}. If there is a trajectory point where the minimum of all gi is > 0, the system may enter the unsafe set. Otherwise there is no such trajectory, and that minimum value is a safety margin.

Demonstrated on a rotated half-circle. Was very annoying trigonometry to get the plots working.