Attempting to write the sampler
what's the deal with deal? Trying to use the event handler in ode45 by allowing the event to be an anonymous function.

Perhaps this more trouble than it is worth, and there exists a better solution.

Right now, matlab is reporting that the number of inputs and outputs disagree. I don't know what to do with that information.