Started writing peak_estimate_dual
Generalized code for peak estimation on SOS (dual) problem.

This uses YALMIP/MOSEK to solve the problem. There is no certification of (near)-optimality nor extraction of moments, but the numerical accuracy is higher as compared to gloptipoly.

Right now, the outputs of peak_estimate_dual_test and peak_dual_test are disagreeing, even for the same problem. This will be the subject of the next investigation.