Pendulum optimality
There's a symmetry in the pendulum system: (c, s, w) <> (c, -s, -w). This is why the moment solution is rank2.

By restricting to positive vangles with s >= 0, the symmetry is broken, and optimality is validated. This is forced by the HALF_ANGLE option in pendulum_test.m

Made some plots for a report