polyval_func is unstable
In polynomials such as the switching-system, yalmip reinterprets some polynomials as BMIs, and adds new variables. This does not affect optimization, but does change function evaluation.

The input should be written as syms, and then somehow converted to yalmip. That would be the easiest transformation. yalmip->syms is hard. And then syms may be used for evaluation maybe?