Compared against safety-time occupation measure system
This is a competing paper (safety_time_occupation.pdf) that upper bounds the maximal amount of time trajectories spend in the unsafe set Xu. Due to numerical inaccuracies (time must be exactly 0), it is subpar for proving avoidance of sets.  Bounds for the flow safety margin system were computed in safety_time_flow.m