Fixed the ordering issue in mmon
Previously, the order of monomials generated by mmon depended on the order in which those variables were declared. If variables were declared in the order of [a, b, c], then monom([c,a,b], degree) results in (1, a, b, c, a^2, ....) instead of (1, c, a, b, c^2 ...). This is because @mpol/indvar sorts variables.

Also did some more tests on the flow model.

Next step: discrete sampler